+++ tests/schema/samples/storeItem/ts-extended.ts
@@ -1,6 +1,7 @@
-import { Summary, Schema } from '../../../../dist'
+import { Schema } from '../../../../dist'
 import { KeyOf } from '../../../../src/schema/types'
-import { IValidationError } from '../../../../src/schema/utils'
+import { FieldError, IValidationError } from '../../../../src/schema/utils'
+import { PayloadKey } from '../../../../src/utils'
 
 type Input = {
   name: string
@@ -13,48 +14,56 @@ type Output = {
   age: string
 }
 
-const schema = new Schema<Output, Input, {}, VError<KeyOf<Input>>>({
+const schema = new Schema<Input, Output, {}, VError<KeyOf<Input>>>({
   age: { default: '', dependsOn: 'dob', resolver: () => '' },
   dob: { default: '' },
   name: { default: '' }
 })
 
-const model = schema.getModel()
+// const model = schema.getModel()
 
-// model.create().then(({ error }) => {
-//   error[0].
+// model.create({}).then(({ error }) => {
+//   error?.[0].key
 // })
 
 type EInput = {
   name: string
   dob: string
   lol: string
+  v: string
 }
 
 type EOutput = {
-  name: string
-  lol: string
-  deps: string
+  age: number
+  dependent: string
   dob: string
-  age: string
+  lol: string
+  name: string
 }
 
 const extended = schema
-  .extend<EOutput, EInput>({
+  .extend<EInput, EOutput>({
     lol: { default: '' },
-    deps: { default: '', dependsOn: 'lol', resolver: () => '' }
+    dependent: { default: '', dependsOn: 'lol', resolver: () => '' },
+    v: { virtual: true, validator: () => true }
   })
   .getModel()
 
-// extended.create().then(({ error }) => {
-//   error[0].key=''
-// })
+extended.create().then(({ data }) => {
+  data?.age
+})
 
 type ErrorData<Keys> = { key: Keys; message: string }[]
 
 class VError<Keys> extends Error implements IValidationError<ErrorData<Keys>> {
   constructor(public message: IValidationError<ErrorData<Keys>>['message']) {
     super(message)
+  }
+  add(
+    field: PayloadKey,
+    value?: string | string[] | FieldError | undefined
+  ): this {
+    throw new Error('Method not implemented.')
   }
 
   get data(): ErrorData<Keys> {